+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Notifications and Requests:
 
 Types:
 - Renamed `TextDocumentPosition` to `TextDocumentPositionParams`.
+- Don't implement `TextDocumentIdentifier` on `FileEvent`.
 - `TextDocumentPositionParams` -- Replace the `uri` parameter with `textDocument`.
 
 ## [1.0.1] - 2023-05-16